requirements.txt
django>=1.6
django-reversion
psycopg2
